requirements.txt
argparse
pathlib
pyyaml